refactor: move everything temporarily into single file so it's easier to purge
