+++ src/index.ts
@@ -1,15 +1,1 @@
-// Main package exports
-export {reactiveComponent, reactiveComponent as rxComponent, forwardRef} from './reactiveComponent'
-export {observeState, observeCallback, observeContext, observeElement} from './utils'
-export {
-  observeState as state,
-  observeCallback as handler,
-  observeContext as context,
-  observeElement as elementRef,
-} from './utils'
-export {observableCallback} from 'observable-callback'
-
-// hooks
-export {useObservable, useMemoObservable} from './useObservable'
-export {useAsObservable} from './useAsObservable'
-export {useObservableCallback} from './useObservableCallback'
+export * from './scratchpad'
